% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadmfuzz.R
\name{compare_gadm_entrie_with_region_names}
\alias{compare_gadm_entrie_with_region_names}
\title{For a given entry perform fuzzy string matching against all subregion names}
\usage{
compare_gadm_entrie_with_region_names(gadm_entrie, region_names)
}
\arguments{
\item{gadm_entrie}{[character vector]}

\item{region_names}{}
}
\value{
A data.frame of with number of rows equals to
`length(gadm_entrie) x length(region_names)` containing the fuzzy string matching
scores.
}
\description{
For a given entry perform fuzzy string matching against all subregion names
}
