% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gadm.R
\name{download_gadm}
\alias{download_gadm}
\title{Download GADM country border polygons}
\usage{
download_gadm(iso3, path, level = 0)
}
\arguments{
\item{iso3}{A character, iso3c code for the country of interest.}

\item{path}{A character string, path of the directory where files should be saved.}

\item{level}{A integer (0, 1, or 2), corresponding to the gadm level:
0 corresponds to country borders while 1 and 2 correspond to administrative units
within the country:
\itemize{
\item 1 to biggest administrative units for a given country.
\item 2 to smallest available administrative units for a given country.
}}
}
\description{
Download GADM country border polygons, by default at highest scale possible,
level 0. It is also possible to specify at which gadm level the polygons should
correspond to.
}
