% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gadm36_EGY_1_sf}
\alias{gadm36_EGY_1_sf}
\title{GADM Polygons for Egypt Governorates}
\format{
A sf object.
}
\source{
https://biogeo.ucdavis.edu/data/gadm3.6/Rsf/gadm36_EGY_1_sf.rds
}
\usage{
gadm36_EGY_1_sf
}
\description{
This data set holds the GADM polygons of Egypt Governorates. This is
used to demonstrate the usage of \code{gadmfuzz::find_best_match()}.
}
\keyword{datasets}
