% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wheat_yield_EGY}
\alias{wheat_yield_EGY}
\title{Dummy wheat yields for Egypt Governorates}
\format{
A data.frame with 2 columns and 27 rows.
}
\usage{
wheat_yield_EGY
}
\description{
This data set holds dummy data for wheat yield in each Egypt governorate.
The column \emph{governorate} holds Governorate names following arbitrary English
spellings that are used to demonstrate the usage of \code{gadmfuzz::find_best_match()}.
}
\keyword{datasets}
