% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadmfuzz.R
\name{determine_level_gadm}
\alias{determine_level_gadm}
\title{Identify level of gadm polygon}
\usage{
determine_level_gadm(gadm_country)
}
\arguments{
\item{gadm_country}{[sf object] gadm polygon geometries for a given country.}
}
\value{
[integer] 0, 1 or 2
}
\description{
Identify level of gadm polygon
}
