% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadmfuzz.R
\name{get_all_english_spellings}
\alias{get_all_english_spellings}
\title{Get all english spellings of a given gadm polygon regions}
\usage{
get_all_english_spellings(gadm_country)
}
\arguments{
\item{gadm_country}{A sf object holding geometries for a given gadm country.}
}
\description{
Get all english spellings of a given gadm polygon regions
}
